+++ src/init.c
@@ -66,6 +66,7 @@ extern SEXP vctrs_as_df_row(SEXP, SEXP);
 extern SEXP vctrs_outer_names(SEXP, SEXP, SEXP);
 extern SEXP vctrs_df_size(SEXP);
 extern SEXP vctrs_as_df_col(SEXP, SEXP);
+extern SEXP vctrs_apply_name_spec(SEXP, SEXP, SEXP, SEXP);
 
 // Defined below
 SEXP vctrs_init(SEXP);
@@ -130,6 +131,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_outer_names",                (DL_FUNC) &vctrs_outer_names, 3},
   {"vctrs_df_size",                    (DL_FUNC) &vctrs_df_size, 1},
   {"vctrs_as_df_col",                  (DL_FUNC) &vctrs_as_df_col, 2},
+  {"vctrs_apply_name_spec",            (DL_FUNC) &vctrs_apply_name_spec, 4},
   {NULL, NULL, 0}
 };
 
@@ -146,7 +148,7 @@ static const R_ExternalMethodDef ExtEntries[] = {
   {"vctrs_cast_common",                (DL_FUNC) &vctrs_cast_common, 1},
   {"vctrs_rbind",                      (DL_FUNC) &vctrs_rbind, 3},
   {"vctrs_cbind",                      (DL_FUNC) &vctrs_cbind, 3},
-  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 2},
+  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 3},
   {NULL, NULL, 0}
 };